+++ src/main/java/com/example/vocatest/controller/MainController.java
@@ -1,0 +1,9 @@
+package com.example.vocatest.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MainController {
+
+}
