+++ src/main/java/com/example/vocatest/dto/UserDto.java
@@ -1,0 +1,52 @@
+package com.example.vocatest.dto;
+
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Map;
+
+public class UserDto implements OAuth2User {
+
+    private final OAuth2Response oAuth2Response;
+
+    private final String role;
+
+    public UserDto(OAuth2Response oAuth2Response, String role) {
+        this.oAuth2Response = oAuth2Response;
+        this.role =role;
+    }
+
+    @Override
+    public Map<String, Object> getAttributes() {
+        return oAuth2Response.getAttributes();
+    }
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+
+        Collection<GrantedAuthority> collection = new ArrayList<>();
+
+        collection.add(new GrantedAuthority() {
+
+            @Override
+            public String getAuthority() {
+
+                return role;
+            }
+        });
+
+        return collection;
+    }
+
+    @Override
+    public String getName() {
+
+        return oAuth2Response.getName();
+    }
+
+    public String getUsername() {
+
+        return oAuth2Response.getProvider()+" "+oAuth2Response.getProviderId();
+    }
+}
