+++ src/main/java/com/example/vocatest/repository/UserRepository.java
@@ -1,0 +1,11 @@
+package com.example.vocatest.repository;
+
+import com.example.vocatest.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+
+    UserEntity findByUsername(String username);
+    UserEntity findByemail(String email); 
+}
+
