--- vocatest/src/main/java/com/example/vocatest/config/SecurityConfig.java
@@ -1,45 +1,0 @@
-package com.example.vocatest.config;
-
-import com.example.vocatest.service.UserService;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration 
-@EnableWebSecurity 
-public class SecurityConfig {
-
-    private final UserService userService;
-
-    public SecurityConfig(UserService custom0Auth2UserService){
-        this.userService = custom0Auth2UserService;
-    }
-
-    @Bean 
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-
-        http
-                .csrf((csrf) -> csrf.disable());
-
-        http
-                .formLogin((login) -> login.disable()); 
-
-        http
-                .httpBasic((basic) -> basic.disable());
-
-        http
-                .oauth2Login(oauth2 -> oauth2
-                        .userInfoEndpoint(userInfoEndpointConfig -> userInfoEndpointConfig
-                                .userService(userService)));
-
-
-        http
-                .authorizeHttpRequests((auth) -> auth
-                        .requestMatchers("/", "/oauth2/**", "/login/**").permitAll()
-                        .anyRequest().permitAll());
-
-        return http.build(); 
-    }
-}
