--- vocatest/src/main/java/com/example/vocatest/controller/MainController.java
@@ -1,9 +1,0 @@
-package com.example.vocatest.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MainController {
-
-}
