--- vocatest/src/main/java/com/example/vocatest/dto/VocaListDto.java
@@ -1,21 +1,0 @@
-package com.example.vocatest.dto;
-
-import com.example.vocatest.entity.VocaListEntity;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class VocaListDto { 
-    private String author;
-    private String title;
-
-    /*
-    public VocaListEntity toEntity(){
-        return new VocaListEntity(author, title); 
-    }
-    */
-        public VocaListEntity createVocalistToEntity(String email){
-        return new VocaListEntity(email, title, 0, 0L);
-    }
-}
