--- vocatest/src/main/resources/static/index.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>OAuth 2.0 로그인</title>
-</head>
-<body>
-<a href="/oauth2/authorization/google" class="btn btn-sm btn-success active" role="button">Google Login</a><br>
-</body>
-</html>
